+++ mobile/android/android-components/components/browser/engine-gecko-beta/src/main/java/mozilla/components/browser/engine/gecko/TrackingProtectionExceptionFileStorage.kt
@@ -103,10 +103,10 @@ internal class TrackingProtectionExceptionFileStorage(
         persist()
     }
 
-    override fun removeAll(activeSessions: List<EngineSession?>?) {
+    override fun removeAll(activeSessions: List<EngineSession>?) {
         runtime.contentBlockingController.clearExceptionList()
-        activeSessions?.forEach {
-            it?.notifyObservers {
+        activeSessions?.forEach { engineSession ->
+            engineSession.notifyObservers {
                 onExcludedOnTrackingProtectionChange(false)
             }
         }